% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-typed.R
\name{assign_typed}
\alias{assign_typed}
\title{Typed Assignment (inline)}
\usage{
assign_typed(sym, call, env = rlang::caller_env())
}
\arguments{
\item{sym}{A symbol}

\item{call}{A call}

\item{env}{An environment}
}
\description{
Typed Assignment (inline)
}
