% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{assigned}
\alias{assigned}
\title{Mark a function argument as potentially requiring type coercion}
\usage{
assigned(x)
}
\arguments{
\item{x}{\verb{[call]}

A call to either:
\itemize{
\item A type-checking function
\item Other argument modifiers (e.g. \code{\link[=required]{required()}}, \code{\link[=optional]{optional()}}, \code{\link[=maybe]{maybe()}})
}}
}
\value{
If used outside of a function typing context (e.g. \code{\link{\%<~\%}} or \code{\link[=typed]{typed()}})
this function raises an error. In a function typing context, returns an
expression marked as being an argument to re-assign during type-checking.
}
\description{
Indicates that a function argument should be assigned to the value returned
by it's type-checking function. This is useful for when a type-checking
function coerces it's input to the correct type.

This function may only be used when declaring a function with \code{\link{\%<~\%}}
or \code{\link[=typed]{typed()}}.
}
\examples{
# Define a check for whether an argument is factor-like
check_factorish <- function(x, x_name = rlang::caller_arg(x)) {
  if (is.factor(x) || is.character(x)) {
    return(as.factor(x))
  }
  rlang::abort(sprintf("`\%s` must be a factor or a character.", x_name))
}

# Retrieve the unique levels of a factor-like vector
levels2 <- typed(function(x = assigned(check_factorish())) {
  base::levels(x)
})
levels2(factor(c("A", "B", "C")))
levels2(c("A", "B", "C"))

# `levels2()` relies on `check_factorish()` to coerce `x` into
# a factor variable. If we don't assign the result of `check_factorish(x)`
# to the argument `x`, then we get an unexpected result.
levels2_bad <- typed(function(x = check_factorish()) {
  base::levels(x)
})
levels2_bad(c("A", "B", "C"))
}
