% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{base_model}
\alias{base_model}
\title{Create a model factory function}
\usage{
base_model(
  fields = list(),
  ...,
  .model_config = model_config(),
  .validators_before = list(),
  .validators_after = list()
)
}
\arguments{
\item{fields}{description}

\item{...}{description}

\item{.model_config}{description}

\item{.validators_before}{description}

\item{.validators_after}{description}
}
\value{
A model factory function.
}
\description{
Create a model factory function
}
