% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{check_args}
\alias{check_args}
\title{Check function arguments}
\usage{
check_args(...)
}
\arguments{
\item{...}{description}
}
\value{
The caller environment.
}
\description{
Check function arguments
}
