% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-typed.R
\name{check_missing_assignment}
\alias{check_missing_assignment}
\title{Check that an object is not missing}
\usage{
check_missing_assignment(expr, name, error_call)
}
\arguments{
\item{expr}{\verb{[expression]}

An expression.}

\item{name}{\verb{[character(1)]}

An object name.}

\item{error_call}{\verb{[environment]}

An environment.}
}
\value{
An error if \code{expr} is missing and \code{NULL} otherwise.
}
\description{
This function is used internally by the active binding functions of objects
typed using \link{\%<~\%} or \code{\link[=assign_typed]{assign_typed()}}. It is not meant for use outside of
this context.
}
\examples{
foo <- function(x) {
  check_missing_assignment(
    expr = x,
    name = "x",
    error_call = rlang::caller_env()
  )
}
try(foo())
}
