% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed-function.R
\name{check_required_arg}
\alias{check_required_arg}
\title{Check that a typed argument is supplied}
\usage{
check_required_arg(x)
}
\arguments{
\item{x}{A function argument.}
}
\value{
An error if \code{x} is missing and \code{NULL} otherwise.
}
\description{
This function is inserted into the body of \code{typed()} functions which
have \code{required()} arguments. It is not meant to be used outside of this
context.
}
\examples{
foo <- function(x) { check_required_arg(x) }
try(foo())
}
