% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{check_type_alias_dots}
\alias{check_type_alias_dots}
\title{Check that one unnamed argument is supplied to dots}
\usage{
check_type_alias_dots(...)
}
\arguments{
\item{...}{An unnamed argument.}
}
\value{
An error if too many dots or an unnamed dot is provided and \code{NULL} otherwise.
}
\description{
This function is used internally by functions generated via \code{type_alias()}.
It is not meant to be used outside of this context.
}
\examples{
foo <- function(...) { check_type_alias_dots(...) }
try(foo(10, 11))
}
