% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{either}
\alias{either}
\title{Allow multiple types}
\usage{
either(...)
}
\arguments{
\item{...}{Type check functions or type strings}
}
\value{
A type check function
}
\description{
Allow multiple types
}
\examples{
m <- typed_struct(
  a = Union(is.integer, is.null),
  b = Union("integer", "logical")
)

# Succeeds
m(a = 10L, b = TRUE)

# Also succeeds
m(a = NULL, b = 10L)
}
