% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed-function.R
\name{has_typed_args}
\alias{has_typed_args}
\title{Test if a function has typed arguments}
\usage{
has_typed_args(x)
}
\arguments{
\item{x}{\verb{[function]}

A function to test.}
}
\value{
\code{TRUE} if \code{x} has typed arguments, \code{FALSE} otherwise.
}
\description{
This function returns \code{TRUE} for typed functions (class \code{"typewriter_typed_function"})
which have any typed arguments and returns \code{FALSE} if \code{x} is a function with
no typed arguments.
}
\examples{
foo <- typed(function(x) { x })
bar <- typed(function(x = rlang::check_required()) { x })
has_typed_args(foo)
has_typed_args(bar)
has_typed_args(base::mean)
}
