% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{is_any}
\alias{is_any}
\title{Type predicate \code{any}}
\usage{
is_any(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\value{
\code{TRUE}
}
\description{
Type predicate \code{any}
}
\examples{
my_model <- base_model(
  a = is_any,
  b = is.integer
)

# Succeeds
my_model(a = 10, b = 20L)

# Also succeeds
my_model(a = "Hi", b = 20L)

}
