% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{is_not_missing_or}
\alias{is_not_missing_or}
\title{Execute code only if a variable is missing}
\usage{
is_not_missing_or(name, or, envir = parent.frame())
}
\arguments{
\item{name}{A value to test for missingings}

\item{or}{An expression to evaluate if \code{name} is missing}

\item{envir}{The environment in which to check for \code{name}}
}
\value{
The value of \code{name} if it is not missing, or \code{NULL} otherwise.
If \code{name} is missing, this function is used for the side-effects produced
by executing \code{or}.
}
\description{
Execute code only if a variable is missing
}
