% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{is_optional}
\alias{is_optional}
\title{Mark a parameter as optional}
\usage{
is_optional(fn)
}
\arguments{
\item{fn}{Type check function}
}
\value{
type check function
}
\description{
Mark a parameter as optional
}
\examples{
my_model <- base_model(
  a = is.integer,
  b = is_optional(is.integer)
)

my_model(a = 10)
}
