% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed-function.R
\name{is_typed_function}
\alias{is_typed_function}
\title{Test if the object is a typed function}
\usage{
is_typed_function(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a typed function, \code{FALSE} otherwise.
}
\description{
This function returns \code{TRUE} for typed functions (class \code{"typewriter_typed_function"})
and returns returns \code{FALSE} otherwise.
}
\examples{
foo <- typed(function(x) { x })
is_typed_function(foo)
is_typed_function(base::mean)

}
