% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-typed.R
\name{is_typed_object}
\alias{is_typed_object}
\title{Test if the object is typed}
\usage{
is_typed_object(x, env = rlang::caller_env())
}
\arguments{
\item{x}{\verb{[object]}

An object provided by name.}

\item{env}{\verb{[environment]}

The environment in which to check whether \code{x} is a typed object. Defaults to
the calling environment of \code{is_typed_object()}.}
}
\value{
\code{TRUE} if \code{x} is a typed object, \code{FALSE} otherwise.
}
\description{
This function returns \code{TRUE} if \code{x} is an object typed by \code{\link{\%<~\%}} or \code{\link[=assign_typed]{assign_typed()}}
in \code{env} and \code{FALSE} otherwise. The object to test must be provided by name.
}
\examples{
\dontshow{if (requireNamespace("chk", quietly = TRUE)) withAutoprint(\{ # examplesIf}
x \%<~\% chk::chk_integer(10L)
y <- 10L
identical(x, y)
is_typed_object(x)
is_typed_object(y)
\dontshow{\}) # examplesIf}
}
