% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{is_typewriter_model}
\alias{is_typewriter_model}
\title{Type predicate \verb{typewriter model}}
\usage{
is_typewriter_model(model_fn)
}
\arguments{
\item{model_fn}{A model factory function created with \code{\link[=base_model]{base_model()}}.}
}
\description{
Type predicate \verb{typewriter model}
}
