% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed-function.R
\name{maybe}
\alias{maybe}
\title{Mark a function argument as being potentially NULL}
\usage{
maybe(x)
}
\arguments{
\item{x}{\verb{[call]}

A call to either:
\itemize{
\item A type-checking function
\item Other argument modifiers (e.g. \code{\link[=static]{static()}}, \code{\link[=assigned]{assigned()}}, \code{\link[=required]{required()}}, \code{\link[=optional]{optional()}})
}}
}
\value{
If used outside of a function typing context (e.g. \code{\link{\%<~\%}} or \code{\link[=typed]{typed()}})
this function raises an error. In a function typing context, returns an
expression marked as being a potentially \code{NULL} argument.
}
\description{
Indicates that a function argument may be \code{NULL} A maybe argument
is only type-checked if it is not \code{NULL}. If a maybe argument is \code{NULL},
then the type check is skipped.

This is useful to accommodate the common pattern of using a \code{NULL}
argument value to indicate that the argument is missing.

This function may only be used when declaring a function with \code{\link{\%<~\%}}
or \code{\link[=typed]{typed()}}.
}
\examples{
\dontshow{if (requireNamespace("chk", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# The `collapse` argument of `base::paste0()` is either
# `NULL` (the default) or a character string. We can use
# `maybe()` to maintain this requirement in a strictly
# typed version of `base::paste0()`.
my_paste0 <- typed(function(
  ... = chk::chk_character(),
  collapse = maybe(chk::chk_string(NULL)),
  recycle0 = chk::chk_flag(FALSE)
) {
  base::paste0(..., collapse = collapse, recycle0 = recycle0)
})
my_paste0(c("A", "B", "C"))
my_paste0(c("A", "B", "C"), collapse = "-")

# Without `maybe()` the default use-case of `my_paste0()` would
# cause an error.
my_paste0_bad <- typed(function(
  ... = chk::chk_character(),
  collapse = chk::chk_string(NULL),
  recycle0 = chk::chk_flag(FALSE)
) {
  base::paste0(..., collapse = collapse, recycle0 = recycle0)
})
try(my_paste0_bad(c("A", "B", "C")))
\dontshow{\}) # examplesIf}
}
