% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{model_config}
\alias{model_config}
\title{Create a model config object}
\usage{
model_config(allow_extra = FALSE, str_to_lower = FALSE, ...)
}
\arguments{
\item{allow_extra}{description}

\item{str_to_lower}{description}
}
\value{
A model config object that can be used in \code{\link[=base_model]{base_model()}}.
}
\description{
Create a model config object
}
