% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{model_dump}
\alias{model_dump}
\title{Convert model to base list}
\usage{
model_dump(obj, by_alias = FALSE, ...)
}
\arguments{
\item{by_alias}{Use aliases for names.}

\item{...}{\strong{not used} at the moment.}
}
\value{
base list object
}
\description{
Convert model to base list
}
