% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{model_field}
\alias{model_field}
\title{Create a model field}
\usage{
model_field(fn, default = NA, alias = NULL, ...)
}
\arguments{
\item{fn}{A type check function.}

\item{default}{A default value for the field.}

\item{alias, ...}{\strong{not used} at the moment}
}
\description{
Create a model field
}
