% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive-model.R
\name{model_from_template}
\alias{model_from_template}
\title{Create a model factory function from a template object}
\usage{
model_from_template(template, use_defaults = FALSE, use_length = FALSE)
}
\arguments{
\item{template}{A template list to derive model from.}

\item{use_defaults}{Whether to use template values as default values.}

\item{use_length}{\strong{not used} at the moment}
}
\value{
A model factory function.
}
\description{
Create a model factory function from a template object
}
\examples{
template <- list(
  a = 10L,
  b = 10L,
  txt = "Some Text"
)

my_model <- model_from_template(template, use_defaults = TRUE)

# Succeeds
my_model(a = 20L)

# Fails
try(my_model(10))
}
