% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{model_validate}
\alias{model_validate}
\title{Validate a list or a data frame}
\usage{
model_validate(obj, model_fn)
}
\arguments{
\item{obj}{A list or a data.frame.}

\item{model_fn}{A model factory function created with \code{\link[=base_model]{base_model()}}.}
}
\description{
Validate a list or a data frame
}
