% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{names_to_camel_case}
\alias{names_to_camel_case}
\title{Convert names of an object to camel case}
\usage{
names_to_camel_case(obj, .recursive = TRUE)
}
\arguments{
\item{obj}{A list or a data frame.}

\item{.recursive}{Whether to convert names on all levels.}
}
\value{
The input object with camel case names.
}
\description{
Convert names of an object to camel case
}
