% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{names_to_snake_case}
\alias{names_to_snake_case}
\title{Convert names of an object to snake case}
\usage{
names_to_snake_case(obj, .recursive = TRUE)
}
\arguments{
\item{obj}{A list or a data frame.}
}
\value{
The input object with snake case names.
}
\description{
Convert names of an object to snake case
}
