% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{optional}
\alias{optional}
\title{Mark a function argument as being optional}
\usage{
optional(x)
}
\arguments{
\item{x}{\verb{[call]}

A call to either:
\itemize{
\item A type-checking function
\item Other argument modifiers (e.g. \code{\link[=static]{static()}}, \code{\link[=assigned]{assigned()}}, \code{\link[=maybe]{maybe()}})
}}
}
\value{
If used outside of a function typing context (e.g. \code{\link{\%<~\%}} or \code{\link[=typed]{typed()}})
this function raises an error. In a function typing context, returns an
expression marked as being an optional argument.
}
\description{
Indicates that a function argument may be unsupplied. An optional argument
is only type-checked if it is supplied. If an optional argument is missing,
then the type check is skipped.

This function may only be used when declaring a function with \code{\link{\%<~\%}}
or \code{\link[=typed]{typed()}}.
}
\examples{
# Define an assertion for an argument being a string
assert_string <- function(x, x_name = rlang::caller_arg(x)) {
  if (!rlang::is_string(x)) {
    rlang::abort(sprintf("`\%s` must be a string.", x_name))
  }
}

# Define a message printing function with a default message
cat_message <- typed(function(msg = optional(assert_string())) {
  if (missing(msg)) {
    cat("~No Message Here~\n")
  } else {
    cat(msg, "\n")
  }
})
cat_message("Hi!")
cat_message()

# Without `optional()` we'd fail the `assert_string()` type-check
cat_message_bad <- typed(function(msg = assert_string()) {
  if (missing(msg)) {
    cat("~No Message Here~\n")
  } else {
    cat(msg, "\n")
  }
})
cat_message_bad("Hi!")
try(cat_message_bad())
}
