% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed-function.R
\name{required}
\alias{required}
\title{Mark a function argument as being required}
\usage{
required(x)
}
\arguments{
\item{x}{\verb{[call]}

A call to either:
\itemize{
\item A type-checking function
\item Other argument modifiers (e.g. \code{\link[=static]{static()}}, \code{\link[=assigned]{assigned()}}, \code{\link[=maybe]{maybe()}})
}}
}
\value{
If used outside of a function typing context (e.g. \code{\link{\%<~\%}} or \code{\link[=typed]{typed()}})
this function raises an error. In a function typing context, returns an
expression marked as being a required argument.
}
\description{
Indicates that a function argument must be supplied. This modifier adds
an additional check which raises an informative error if an argument
with no default value is not supplied when the function is called.

This function may only be used when declaring a function with \code{\link{\%<~\%}}
or \code{\link[=typed]{typed()}}.
}
\details{
When an argument is \code{required()} a call to \code{\link[=check_required_arg]{check_required_arg()}} is inserted
into the typed function's body before the argument's type is checked.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Typed function declaration
foo <- typed(function(x = required(chk::chk_integer())) \{
  return(x)
\})

# Roughly equivalent to
foo <- function(x) \{
  typewriter::check_required_arg(x)
  chk::chk_integer(x)
  return(x)
\}
}\if{html}{\out{</div>}}
}
\examples{
# Define an assertion for an argument being numeric
assert_numeric <- function(x, x_name = rlang::caller_arg(x)) {
  if (!is.numeric(x)) {
    rlang::abort(sprintf("`\%s` must be numeric.", x_name))
  }
}

# Define a function with numeric-typed argument `x`
double <- typed(function(x = assert_numeric()) {
  x * 2
})

# If `x` is un-supplied, we get an error from `assert_numeric()`,
# not `double()`
try(double())

# We can correct this with `required()`
double2 <- typed(function(x = required(assert_numeric())) {
  x * 2
})
try(double2())
}
