% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed-function.R
\name{rethrow_parent_assignment_error}
\alias{rethrow_parent_assignment_error}
\title{Re-signal an assignment error}
\usage{
rethrow_parent_assignment_error(expr)
}
\arguments{
\item{expr}{\verb{[expression]}

An expression.}
}
\value{
The result of evaluating \code{expr} or an error.
}
\description{
This function is inserted into the body of \code{typed()} functions which
have \code{static()} arguments. It is not meant to be used outside of this
context.
}
\examples{
try(rethrow_parent_assignment_error(x \%<~\% stop("Parent Error")))
}
