% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_syntax.R
\name{split_fn_type_checking}
\alias{split_fn_type_checking}
\title{Split arguments to extract type information}
\usage{
split_fn_type_checking(fn)
}
\arguments{
\item{fn}{A function definition from which to extract parameter information}
}
\value{
A list of split type information as provided by
\code{\link{split_arg_type_checking}}
}
\description{
Split arguments to extract type information
}
\seealso{
Other type_parsing: 
\code{\link{split_arg_type_checking}()}
}
\concept{type_parsing}
