% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{type_alias}
\alias{type_alias}
\title{Create a type alias}
\usage{
type_alias(call, name = NULL, env = caller_env(), returns = c("arg", "result"))
}
\arguments{
\item{call}{A call}

\item{name}{A string}

\item{env}{An environment}

\item{returns}{A result}
}
\value{
An alias
}
\description{
Create a type alias
}
\examples{
type_alias(foo)

}
