% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_check.R
\name{type_check}
\alias{type_check}
\title{Type check named arguments against type paramvalsraints}
\usage{
type_check(..., type_params)
}
\arguments{
\item{...}{Named type definition pairs. The name of the parameter represents
the variable in the parent environment which should be type checked, and
the value represents the type definition.}

\item{type_defs}{A \code{character} vector of type parameters}
}
\value{
A list of type bounds and their observed values over tested variables
}
\description{
Type check named arguments against type paramvalsraints
}
\seealso{
Other type-evaluation: 
\code{\link{\%is\%}()},
\code{\link{type_constrain}()}
}
\concept{type-evaluation}
