% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_constrain.R
\name{type_constrain}
\alias{type_constrain}
\title{Apply type constraints over observed type parameter values}
\usage{
type_constrain(paramvals, ...)
}
\arguments{
\item{paramvals}{A list of type parameters and their observed values}

\item{...}{Type constraint expressions, from which constraint functions can
be produced.}
}
\value{
Returns \code{NULL}, though this function is called primarily for its
side-effects of issueing errors when type constraint checking fails.
}
\description{
Apply type constraints over observed type parameter values
}
\seealso{
Other type-evaluation: 
\code{\link{\%is\%}()},
\code{\link{type_check}()}
}
\concept{type-evaluation}
