% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed-struct.R
\name{typed_struct}
\alias{typed_struct}
\title{Create a typed structure}
\usage{
typed_struct(...)
}
\arguments{
\item{...}{Type definitions (Type check functions)}
}
\description{
Create a typed structure
}
