% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Union}
\alias{Union}
\title{Allow multiple types}
\usage{
Union(...)
}
\arguments{
\item{...}{Type check functions}
}
\value{
A type check function
}
\description{
Allow multiple types
}
\examples{
{
  m <- base_model(
    a = Union(is.integer, is.null)
  )
  m(10L)
  m(NULL)
}
}
