% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed-function.R
\name{untype_function}
\alias{untype_function}
\title{Unclass a typed function}
\usage{
untype_function(x)
}
\arguments{
\item{x}{\verb{[function]}

A typed or regular function to un-type.}
}
\value{
A function.
}
\description{
Converts a typed function (class \code{"typewriter_typed_function"}) into a regular
function. This removes any type-checking calls added prior to the typed
function's body and sets the function's class back to \code{function}. If \code{x} is
an untyped function, then \code{x} is returned.
}
\examples{
# Define a type-checking function
assert_bool <- function(x, x_name = rlang::caller_arg(x)) {
  if (!rlang::is_bool(x)) {
    rlang::abort(sprintf("`\%s` must be `TRUE` or `FALSE`.", x_name))
  }
}

# `base::sum()` doesn't check the type of `na.rm`
sum2 <- typed(function(..., na.rm = assert_bool(FALSE)) {
  base::sum(..., na.rm = na.rm)
})
try(sum2(1, 2, NA, na.rm = "NO!"))

# Remove type checking on `na.rm`
sum2 <- untype_function(sum2)
sum2(1, 2, NA, na.rm = "NO!")
}
