% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{untyped}
\alias{untyped}
\title{Mark a function argument as untyped}
\usage{
untyped(x)
}
\arguments{
\item{x}{\verb{[expression]}

An expression to use as the default argument value.}
}
\value{
If used outside of a function typing context (e.g. \code{\link{\%<~\%}} or \code{\link[=typed]{typed()}})
this function raises an error. In a function typing context, returns an
expression marked as being an untyped argument.
}
\description{
Indicates that a typed function argument should be considered untyped.
This function may only be used when declaring a function with \code{\link{\%<~\%}}
or \code{\link[=typed]{typed()}}. This is useful for preventing the default value of an
argument from being considered a type-checking call.
}
\examples{
# `typed()` and `\%<~\%` interpret any call as a type call. Here,
# `foo()` attempts to type-check `x` using `as.Date(x)`.
foo <- typed(function(x = as.Date("2020-01-01")) { x })
print(foo)

# To use `as.Date("2020-01-01")` as the default value of `x`,
# you can explicitly mark it as untyped.
foo <- typed(function(x = untyped(as.Date("2020-01-01"))) { x })
print(foo)
}
