% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_parameters.R
\name{which_type_parameters}
\alias{which_type_parameters}
\title{Extract type parameters}
\usage{
which_type_parameters(exprs)
}
\arguments{
\item{exprs}{A list of language objects}
}
\value{
A character vector of names
}
\description{
Genericly, this function simply extract single names from a list of
expressions, returned as a character vector.
}
