% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{check_args}
\alias{check_args}
\title{Check function arguments}
\usage{
check_args(...)
}
\arguments{
\item{...}{Arg definitions.}
}
\value{
The caller environment.
}
\description{
Check function arguments
}
\examples{
{
  f <- function(a, b) {
    check_args(a = is.integer, b = is.integer)
    a + b
  }

  # Succeeds
  f(10L, 20L)

  # Fails
  try(f(10L, 4.6))
}
}
