% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{dtype}
\alias{dtype}
\title{Create a type check function}
\usage{
dtype(type_check, default = NA)
}
\arguments{
\item{type_check}{Type check function or type string.}

\item{default}{A default value.}
}
\value{
A type check function
}
\description{
Create a type check function
}
