% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{dtype}
\alias{dtype}
\title{Create a model field/type definition}
\usage{
dtype(x, default = NA)
}
\arguments{
\item{x}{A type check function or type string.}

\item{default}{A default value.}
}
\value{
A \code{\link[=model_field]{model_field()}}.
}
\description{
Create a model field/type definition
}
