% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{model_config}
\alias{model_config}
\title{Create a model config object}
\usage{
model_config(allow_extra = FALSE, str_to_lower = FALSE, ...)
}
\arguments{
\item{allow_extra}{Whether to allow extra fields without type check.}

\item{str_to_lower}{Convert all strings to lower case.}

\item{...}{\strong{not used} at the moment}
}
\value{
A model config object that can be used in \code{\link[=base_model]{base_model()}}.
}
\description{
Create a model config object
}
\examples{
cfg <- model_config(str_to_lower = TRUE)

my_model <- base_model(
  bar = is.character,
  foo = is.character,
  .model_config = cfg
)

my_model(bar = "FOO", foo = "BAR")
}
