% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{optional}
\alias{optional}
\title{Mark a parameter as optional}
\usage{
optional(obj)
}
\arguments{
\item{obj}{A type definition object.}
}
\value{
A \code{\link[=model_field]{model_field()}} with \code{optional = TRUE}.
}
\description{
Mark a parameter as optional
}
