% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Optional}
\alias{Optional}
\title{Mark a parameter as optional}
\usage{
Optional(type_check_fn)
}
\arguments{
\item{type_check_fn}{Type check function or type string.}
}
\value{
type check function
}
\description{
Mark a parameter as optional
}
\examples{
my_model <- base_model(
  a = "integer",
  b = Optional("integer")
)

my_model(a = 10L)
}
