% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed-struct.R
\name{typed_struct}
\alias{typed_struct}
\title{Create a typed structure}
\usage{
typed_struct(..., .allow_na = FALSE)
}
\arguments{
\item{...}{Type definitions.}

\item{.allow_na}{Allow values to be initialized with \code{NA}.}
}
\value{
A type factory function.
}
\description{
Create a typed structure
}
\examples{
Person <- typed_struct(
  name = "character",
  age = "integer"
)

hanna <- Person(
  name = "Hanna",
  age = 10L
)

hanna

inherits(hanna, "Person")
inherits(hanna, "list")
}
