% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typewriter-package.R
\docType{package}
\name{typewriter-package}
\alias{typewriter}
\alias{typewriter-package}
\title{typewriter: A minimal object typing system for R}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EthanSansom/typewriter}
  \item Report bugs at \url{https://github.com/EthanSansom/typewriter/issues}
}

}
\author{
\strong{Maintainer}: Ethan Sansom \email{ethan.sansom29@gmail.com} (\href{https://orcid.org/0009-0000-1573-0186}{ORCID}) [copyright holder]

}
\keyword{internal}
